Fix for https://github.com/bobbingwide/oik-plugins/issues/1
In oikp_display_documentation only use the value of the _oik_doc_home post meta
if the field is registered for the oik-plugins post type
and the ID contained is a valid page.
This is used to create a [bw_tree] of documentation.

Otherwise build a list of posts and pages with a matching _plugin_ref.